document-analyzer/requirements.txt
fastapi
uvicorn
groq
pymupdf
python-docx
python-dotenv
pydantic
python-multipart
python-pptx
